+++ plugins/Moodle/client.php
@@ -5,8 +5,13 @@
  * XMLRPC client for Moodle 2
  */
 
-// Load cURL class
+// Load cURL class.
 require_once 'classes/curl.php';
+
+// @since 7.6 PHP8 no xmlrpc ext: load XML_RPC compat classes.
+require_once 'plugins/Moodle/includes/XML_RPC.php';
+require_once 'plugins/Moodle/includes/XmlrpcDecoder.php';
+require_once 'plugins/Moodle/includes/XmlrpcEncoder.php';
 
 
 /**
@@ -28,30 +33,32 @@ function moodle_xmlrpc_call( $functionname, $object )
 	//var_dump($object);
 
 	if ( empty( $object ) )
+	{
 		return false;
+	}
 
-	$post = xmlrpc_encode_request(
+	$post = XML_RPC::EncodeRequest(
 		$functionname,
 		$object,
 		array( 'encoding' => 'utf-8', 'escaping' => 'markup' )
 	);
 
-	$resp = xmlrpc_decode( $curl->post( $serverurl, $post ), 'utf-8' );
+	$resp = XML_RPC::Decode( $curl->post( $serverurl, $post ), 'utf-8' );
 
 	if ( get_xmlrpc_error( $resp ) )
 	{
 		// Handle the positive response.
 		return call_user_func( $functionname . '_response', $resp );
 	}
-	else
-		return false;
+
+	return false;
 }
 
 /**
  * Get the XML RPC error if any
  * Adds the error message to the global $error variable
  *
- * @param  string|array $resp cURL POST response
+ * @param  string|array $resp cURL POST response.
  *
  * @return bool               false on error, else true
  */
@@ -60,7 +67,7 @@ function get_xmlrpc_error( $resp )
 	global $error;
 
 	if ( is_array( $resp )
-		&& xmlrpc_is_fault( $resp ) )
+		&& XML_RPC::IsFault( $resp ) )
 	{
 		$message = 'Moodle: ' . $resp['faultCode'] . ' - ' . $resp['faultString'];